Allow host port to be overriden using env var
This change updates the Vagrant file to enable the ability to specify which port should be bound on the host using an environment variable. This gives more flexibility and means a change in code is not required should you need to use a different port number.

The changes also cover updating the README to include details about this ability.