Remove always run start section from Vagrantfile
On another project found that we could infact add a service to start up using update-rc.d (`update-rc.d redis-server defaults`). However upon trying it for Redis got a message back saying this had already been done.

This meant that the final section in the Vagrantfile is no longer necessary so have removed it.